add Senko debug version
