--- happyboard/20250219V2_test_oop__timer/senko copy 4.py
@@ -1,199 +1,0 @@
-import urequests
-import uhashlib
-import gc
-from time import sleep
-import micropython
-
-class Senko:
-    raw = "https://raw.githubusercontent.com"
-    github = "https://github.com"
-
-    def __init__(self, user, repo, url=None, branch="master", working_dir="app", files=["boot.py", "main.py"], headers={}):
-        """Senko OTA agent class.
-
-        Args:
-            user (str): GitHub user.
-            repo (str): GitHub repo to fetch.
-            branch (str): GitHub repo branch. (master)
-            working_dir (str): Directory inside GitHub repo where the micropython app is.
-            url (str): URL to root directory.
-            files (list): Files included in OTA update.
-            headers (list, optional): Headers for urequests.
-        """
-        self.base_url = "{}/{}/{}".format(self.raw, user, repo) if user else url.replace(self.github, self.raw)
-        self.url = url if url is not None else "{}/{}/{}".format(self.base_url, branch, working_dir)
-        self.headers = headers
-        self.files = files
-    # 比對檔案內容
-    def _check_hash(self, x, y):
-        #計算 x（GitHub 版本）與 y（本地版本）的 SHA1 雜湊
-        x_hash = uhashlib.sha1(x.encode())
-        y_hash = uhashlib.sha1(y.encode())
-
-        x = x_hash.digest()
-        y = y_hash.digest()
-
-        if str(x) == str(y):
-            return True #比對 SHA1 是否相同，如果不同代表程式碼有變更，需要更新
-        else:
-            return False
-
-    def _get_file(self, url):
-        # =======
-        # 記憶體測試
-        # =====
-        gc.collect()
-        micropython.mem_info()
-        # 發送 HTTP GET 請求 下載 GitHub 上的最新版本程式碼
-        payload = urequests.get(url, headers=self.headers)
-        code = payload.status_code # 取的狀態碼
-        #print("read ok  "+url)
-        gc.collect()
-        #print(gc.mem_free())
-        if code == 200:
-            return payload.text #回傳文字內容（程式碼）
-        else:
-            return None
-
-    ## 檢查哪些檔案需要更新
-    def _check_all(self):
-        changes = []
-        index = 0
-        print(f"debug:[_check_all]: 進入迴圈前")
-        micropython.mem_info()
-        for file in self.files:
-            # =====
-            # 新版待測試 所以使用60000 總體記憶體為標準試看看
-            # =====
-            index += 1
-            print(f"debug:[_check_all]: 有{file}, {index}")
-
-            while(gc.mem_free() < 60000): #確保記憶體夠用 (gc.mem_free() < 60000
-                # =====
-                # 打印整體記憶體資訊
-                # =====
-                gc.collect()
-                print(f"debug:[_check_all]: 迴圈內")
-                micropython.mem_info()
-                
-                sleep(1)
-            latest_version = self._get_file(self.url + "/" + file)  # 調用方法發送請求 下載 GitHub 上的最新版本程式碼    
-            if latest_version is None:
-                continue
-
-            try:
-                with open(file, "r") as local_file: #讀取 ESP32 本地版本的相同檔案
-                    local_version = local_file.read()
-            except:
-                local_version = ""
-
-            if not self._check_hash(latest_version, local_version):
-                changes.append(file) #如果 不同，代表有變更，加入 changes 清單
-            latest_version=""
-            local_version = ""
-            print(f"debug:[_check_all]: return前")
-            micropython.mem_info()
-        return changes
-
-    # def fetch(self): # 沒有使用
-    #     """Check if newer version is available.
-
-    #     Returns:
-    #         True - if is, False - if not.
-    #     """
-    #     if not self._check_all():
-    #         return False
-    #     else:
-    #         return True
-    # 執行 OTA 更新
-    def update(self):
-        """Replace all changed files with newer one.
-
-        Returns:
-            True - if changes were made, False - if not.
-        """
-        changes = self._check_all() #呼叫 _check_all() 找出需要更新的檔案
-        gc.collect()
-        micropython.mem_info()
-        for file in changes: #逐一下載 GitHub 最新版本並覆蓋 ESP32 上的舊版本
-            with open(file, "w") as local_file:
-                local_file.write(self._get_file(self.url + "/" + file))
-            
-
-        if changes:
-            return True
-        else:
-            return False
-
-
-    # def _get_file(self, url):
-    #     # =======
-    #     # 記憶體測試
-    #     # =====
-    #     gc.collect()
-    #     micropython.mem_info()
-    #     payload = urequests.get(url, headers=self.headers)
-    #     code = payload.status_code
-    #     #print("read ok  "+url)
-    #     gc.collect()
-    #     #print(gc.mem_free())
-    #     if code == 200:
-    #         return payload.text
-    #     else:
-    #         return None
-
-    # def _check_all(self):
-    #     changes = []
-    #     print('debug[senko._check_all] Start') 
-    #     index_times = 0
-    #     print(f"debug[senko._check_all] 進入迴圈前:{gc.mem_free()}")
-
-    #     for file in self.files:
-    #         # =====
-    #         # 測試 
-    #         # =====
-    #         index_times += 1
-    #         print(f'debug:[senko._check_all 迴圈跑到{file}]: index: {index_times}')
-    #         while(gc.mem_free()<60000):
-    #             gc.collect()
-    #             print(f"{file}: ===> (可用記憶體: {gc.mem_free()})")
-         
-    #             sleep(1)
-    #         latest_version = self._get_file(self.url + "/" + file)            
-    #         if latest_version is None:
-    #             continue
-
-    #         try:
-    #             with open(file, "r") as local_file:
-    #                 local_version = local_file.read()
-    #         except:
-    #             local_version = ""
-
-    #         if not self._check_hash(latest_version, local_version):
-    #             changes.append(file)
-    #         latest_version=""
-    #         local_version = ""
-    #         print(f"return之前 ===> (可用記憶體: {gc.mem_free()})")
-    #     return changes
-
-   
-    # def update(self):
-    #     """Replace all changed files with newer one.
-
-    #     Returns:
-    #         True - if changes were made, False - if not.
-    #     """
-    #     changes = self._check_all()
-    #     #print(changes)
-    #     gc.collect()
-    #     #print(gc.mem_free())
-    #     for file in changes:
-    #         with open(file, "w") as local_file:
-    #             local_file.write(self._get_file(self.url + "/" + file))
-            
-
-    #     if changes:
-    #         return True
-    #     else:
-    #         return False
-
