+++ happyboard/20250219V2_test_oop__timer/senko_debug.py
@@ -34,10 +34,28 @@ class Senko:
         x = x_hash.digest()
         y = y_hash.digest()
 
-        if str(x) == str(y):
-            return True #比對 SHA1 是否相同，如果不同代表程式碼有變更，需要更新
-        else:
-            return False
+        ## 取的比對結果
+        result = (str(x) == str(y))
+        # if str(x) == str(y):
+        #     return True #比對 SHA1 是否相同，如果不同代表程式碼有變更，需要更新
+        # else:
+        #     return False
+        print(f"Debugger:[_check_hash] x,y哈希值:{x_hash}，記憶體:{gc.mem_free()}")
+        #回收哈希值
+        del x_hash, y_hash, x, y
+        gc.collect()
+        print(f"Debugger:[_check_hash] 刪除x,y哈希值{result}，記憶體:{gc.mem_free()}")
+        return result
+
+
+        # 這裡的x, y在實際調用時，會傳參 latest_version 和 local_version
+        # 函式裡面又對這兩個參數做x_hash  y_hash 
+        # 但這個.check_hash()函式 我們只需要知道TURE OR FALSE 因此要刪除 x_hash 和y_hash
+        # 後續的gc.collect()才有效果
+        # if str(x) == str(y):
+        #     return True #比對 SHA1 是否相同，如果不同代表程式碼有變更，需要更新
+        # else:
+        #     return False
 
     def _get_file(self, url):
         # =======
@@ -46,14 +64,19 @@ class Senko:
         gc.collect()
         micropython.mem_info()
         # 發送 HTTP GET 請求 下載 GitHub 上的最新版本程式碼
+        print(f"Debugger:[_get_file] 這裡執行.get(): {url}")  # 🔍 加入 URL 印出來檢查
         payload = urequests.get(url, headers=self.headers)
         code = payload.status_code # 取的狀態碼
+        print(f"Debugger:[_get_file] HTTP 狀態碼: {code}") 
         #print("read ok  "+url)
         gc.collect()
         #print(gc.mem_free())
         if code == 200:
-            return payload.text #回傳文字內容（程式碼）
+            data = payload.text
+            print(f"Debugger:[_get_file] status(200)")
+            return data #回傳文字內容（程式碼）
         else:
+            print(f"Debugger:[_get_file] 無法取得 {url}，回傳 None")
             return None
     # def _get_file(self, url):
     #     """ 以 chunked 方式下載檔案，避免記憶體爆滿 """
@@ -94,16 +117,15 @@ class Senko:
             # =====
             index_times += 1
             
-            print(f'Debugger:[senko._check_all 迴圈跑到{file}]: 更新第: {index_times} 檔案')
+            print(f'Debugger:[senko._check_all 迴圈跑到{file}]  第: {index_times} 檔案，記憶體為：')
             micropython.mem_info()
+            ## 記憶體不夠用才會跑這裡 但即使gc後沒有60000 還是出不了這個while??
             while(gc.mem_free() < 60000): #確保記憶體夠用 (gc.mem_free() < 60000
                 # =====
                 # 打印整體記憶體資訊
                 # =====
                 gc.collect()
-                
-                #print(gc.mem_free())
-                print(f"{file}: ===>可用記憶體:")
+                print(f'Debugger:[while內] 釋放記憶體：')
                 micropython.mem_info()
                 sleep(1)
                 
@@ -124,46 +146,28 @@ class Senko:
                 local_version = ""
                 print(f'Debugger:[senko._check_all] {file} 目前local_version為空')
 
+            #### 這裡調用了._check_hash(latest_version, local_version)
             if not self._check_hash(latest_version, local_version):
                 changes.append(file) #如果 不同，代表有變更，加入 changes 清單
                 print(f'Debugger:[senko._check_all] 目前已加入changes清單，有 {changes} ')
-            print(f"Debugger:[latest_version與local_version]尚未是空字串,記憶體:{gc.mem_free()}")
-            latest_version=""
-            local_version = ""
-            print(f"Debugger:[latest_version與local_version]已變成空字串,記憶體:{gc.mem_free()}")
-            ## 
+            else:
+                 # **清空 latest_version 和 local_version**
+                print(f"Debugger:[latest_version與local_version & gc以前] 遠端和本地檔案內容一致，記憶體:{gc.mem_free()}")
+                del latest_version, local_version
+                gc.collect()
+                print(f"Debugger:[latest_version與local_version& gc以後] 遠端和本地檔案內容一致，無須更新，不用重啟，記憶體:{gc.mem_free()}")
             # 釋放記憶體
-            del latest_version, local_version
-            print(f"Debugger:[del後]:{gc.mem_free()}")
             gc.collect()
             print(f"Debugger:[gc後]:{gc.mem_free()}")
         return changes
 
-
-    # def fetch(self): # 沒有使用
-    #     """Check if newer version is available.
-
-    #     Returns:
-    #         True - if is, False - if not.
-    #     """
-    #     if not self._check_all():
-    #         return False
-    #     else:
-    #         return True
     # 執行 OTA 更新
     def update(self):
-        """Replace all changed files with newer one.
-
-        Returns:
-            True - if changes were made, False - if not.
-        """
         changes = self._check_all() #呼叫 _check_all() 找出需要更新的檔案
         gc.collect()
-        # =======
-        # 記憶體打印
-        # =====
-        print('Debugger[senko.update] memory_data') 
-        micropython.mem_info()
+        print(f'Debugger[senko.update] memory_data: {gc.mem_free()}') 
+
+        #這裡指的是 如果有changes回傳 沒有就代表無須更新 遠端檔案 與本地檔案一致
         for file in changes: #逐一下載 GitHub 最新版本並覆蓋 ESP32 上的舊版本
             with open(file, "w") as local_file:
                 local_file.write(self._get_file(self.url + "/" + file))
@@ -172,4 +176,5 @@ class Senko:
         if changes:
             return True
         else:
+            print(f'Debugger[senko.update] 遠端檔案 和本地檔案 一致 無須更新') 
             return False
