+++ happyboard/20250219V2_test_oop__timer/senko_print.py
@@ -1,0 +1,110 @@
+import urequests
+#import uhashlib
+import gc
+from time import sleep
+
+class Senko:
+    raw = "https://raw.githubusercontent.com"
+    github = "https://github.com"
+
+    def __init__(self, user, repo, url=None, branch="master", working_dir="app", files=["boot.py", "main.py"], headers={}):
+        """Senko OTA agent class.
+
+        Args:
+            user (str): GitHub user.
+            repo (str): GitHub repo to fetch.
+            branch (str): GitHub repo branch. (master)
+            working_dir (str): Directory inside GitHub repo where the micropython app is.
+            url (str): URL to root directory.
+            files (list): Files included in OTA update.
+            headers (list, optional): Headers for urequests.
+        """
+        self.base_url = "{}/{}/{}".format(self.raw, user, repo) if user else url.replace(self.github, self.raw)
+        self.url = url if url is not None else "{}/{}/{}".format(self.base_url, branch, working_dir)
+        self.headers = headers
+        self.files = files
+
+    def _check_hash(self, x, y):
+        import uhashlib
+        x_hash = uhashlib.sha1(x.encode())
+        y_hash = uhashlib.sha1(y.encode())
+
+        x = x_hash.digest()
+        y = y_hash.digest()
+
+        if str(x) == str(y):
+            return True
+        else:
+            return False
+
+    def _get_file(self, url):
+        #print(url)
+        gc.collect()
+        #print(gc.mem_free())
+        payload = urequests.get(url, headers=self.headers)
+        code = payload.status_code
+        #print("read ok  "+url)
+        gc.collect()
+        #print(gc.mem_free())
+        if code == 200:
+            return payload.text
+        else:
+            return None
+
+    def _check_all(self):
+        changes = []
+
+        for file in self.files:
+            print(f"Debugger:[while之前]:{gc.mem_free()}")
+            while(gc.mem_free()<60000):
+                gc.collect()
+                print(f"Debugger:[while內清除]:{gc.mem_free()}")
+                sleep(1)
+            latest_version = self._get_file(self.url + "/" + file)            
+            if latest_version is None:
+                continue
+
+            try:
+                with open(file, "r") as local_file:
+                    local_version = local_file.read()
+            except:
+                local_version = ""
+
+            if not self._check_hash(latest_version, local_version):
+                changes.append(file)
+            latest_version=""
+            local_version = ""
+            #print(f"Debugger:[last_line_loop]:{gc.mem_free()}")
+
+        return changes
+
+    def fetch(self):
+        """Check if newer version is available.
+
+        Returns:
+            True - if is, False - if not.
+        """
+        if not self._check_all():
+            return False
+        else:
+            return True
+
+    def update(self):
+        """Replace all changed files with newer one.
+
+        Returns:
+            True - if changes were made, False - if not.
+        """
+        changes = self._check_all()
+        #print(changes)
+        gc.collect()
+        #print(gc.mem_free())
+        for file in changes:
+            with open(file, "w") as local_file:
+                local_file.write(self._get_file(self.url + "/" + file))
+            
+
+        if changes:
+            return True
+        else:
+            return False
