--- happyboard/20250219V2_test_oop__timer/wifi.dat
@@ -1,1 +1,0 @@
-propsky;42886178sky
